--- a/lib/iris/tests/unit/fileformats/um/fast_load_structured_fields/test_BasicFieldCollation.py
+++ b/lib/iris/tests/unit/fileformats/um/fast_load_structured_fields/test_BasicFieldCollation.py
@@ -31,7 +31,8 @@ from iris._lazy_data import as_lazy_data
 from cftime import datetime
 import numpy as np
 
-from iris.fileformats.um._fast_load_structured_fields import FieldCollation
+from iris.fileformats.um._fast_load_structured_fields \
+    import BasicFieldCollation as FieldCollation
 import iris.fileformats.pp