This intends to address problems discussed in [this comment](https://github.com/SciTools/iris/pull/2640#issuecomment-313665559) and follow-ons.

Implementation notes :
  * The information is attached to the 'field' arg in the load callback, i.e. in this case a FieldCollation.
  * The caller should be able to use this to annotate the multidimensional collated cubes as required :  I believe this is what the users's existing code does (using normal load to annotate each raw cube with an extra  "pseudolevel ordering" coordinate).
  * the old 'FieldCollation' class is now renamed 'BasicFieldCollation' : the new, public 'FieldCollation' is a derived class which adds the location-in-input-file info.  This preserves the original, cleaner concept.

Todo:
  * [x] user to confirm suitablility
  * [x] needs a whatsnew