@pp-mo 2018-04-23
> Is it so that the order is round-tripped

Exactly.
They are processing and re-saving data, and need to preserve the original file order for compatibility with existing applications.
To be exact, the input is fieldsfiles, for which field order is supposed to be defined by format specification, but there is an unfortunate loophole relating to 'pseudo-level' encoding, for which the occurrence order is, effectively, significant but not predictable.
Correct me if wrong @hdyson ?

>  technically the order isn't entirely lost ... we store a pointer to the data

That is true, but this enables the purpose to be achieved using public supported API instead of relying on hidden implementation detail.