@hdyson 2018-10-03
> Is this still of interest @hdyson ?

It is, yes - we *need* to be able to preserve the pseudo level order, and would *like* to be able to use the structured loading, so this looks like it hits the sweet spot of giving us the structured loading benefit alongside retaining the existing functionality.