@pp-mo 2018-10-03
Just to paraphrase + review @pelson [previous comments above](https://github.com/SciTools/iris/pull/2977#pullrequestreview-113065457).
I believe I already addressed most of these, as follows ...

  * **What is the usecase for needing the position in the file? Is it so that the order is round-tripped or is there a deeper meaning? ... I think we need to know that we are doing it for good reason ...**
    * the driving usecase seems specific, but the proposal just restores info available to standard load callbacks but missing in structure loads : i.e. which index in which file each field came from

  * **The changes look involved for a use-case that doesn't appear to have been fully clarified. In order to take on such churn, I think we need to know that we are doing it for good reason ...**
    * I honestly don't see that the changes are very involved.  An index property is added to each field; a new collation type stores the filepath, and adds access methods for the new info.

  * **users are going to need to dig deep into the Iris API anyway (to control the save order) ...**
    * not really, you can use the save_pairs interface effectively to do this.

  * **technically the order isn't entirely lost... we store a pointer to the data ...**
    * this is not available via public API


The remaining point is about performance, which I think is worthy of more comment ...

  * **You are touching performance sensitive code. Can you give some reproducible remarks (read: not a full-on performance assessment, more a quick sense of the implications) on what impact the changes have?**

From a performance PoV, nothing has changed in the structure identification step.  PPField gain an extra assignment in `__init__`, and loading -- for structure loading only -- has an additional iterator that assigns a value to each field .  I really can't see anything very costly in this.
The purpose of structured um loading is to avoid : (a) running rules and (b) creating cube objects for each PPField object, and (c) a raw-cube merging stage.
None of that changes, so the only real cost is a small slowdown in creating PPField objects themselves.
This does apply to 'normal' loads as well.  It is in principle avoidable (we don't need to always initialise the private '_index_in_structured_load_file'  property), but I think it is wiser to include the new thing in the `__init__`.  

Here's are some desktop measures of the timing differences, loading from `dataZoo/FF/ff_variants/ff_64.ff`, containing 1202 fields 
  * for making a list of PPFields with `fields = list(iris.fileformats.um.um_to_pp(path))` :
    * old Iris 2.1.0 : (0.533359, 0.528687, 0.527758, 0.556547, 0.529979) ~ 0.5353 seconds
    * new branch code : (0.533923, 0.528155, 0.526929, 0.526727, 0.530310) ~ 0.5292 seconds
  * a 'normal' load with `data = iris.load(path)` :
    * old : (5.352467, 5.332399, 5.365070, 5.373689, 5.333997) ~  5.3515 seconds
    * new : (5.367050, 5.273514, 5.282430, 5.262198, 5.349504) ~ 5.3069 seconds
  * a 'structured' load with `iris.fileformats.structured_loading` and `iris.load` :
    * old : (2.515778, 2.420102, 2.439936, 2.452089, 2.424549) ~  2.4505 seconds
    * new : (2.456725, 2.454413, 2.452883, 2.425538, 2.430701) ~ 2.4441 seconds

Not a great testcase, but nothing much to report.